refactor: move cli fleet modules into cli/fleet/ subpackage
Convert _fleet.py facade to fleet/__init__.py and move _fleet_display,
_fleet_lifecycle, _fleet_session alongside it.
