+++ src/autoskillit/cli/__init__.py
@@ -11,7 +11,7 @@ from pathlib import Path  # noqa: F401 — tests patch autoskillit.cli.Path.home
 
 from autoskillit.cli._cook import cook
 from autoskillit.cli.doctor import DoctorResult
-from autoskillit.cli._fleet import fleet_campaign, fleet_dispatch, fleet_list, fleet_status
+from autoskillit.cli.fleet import fleet_campaign, fleet_dispatch, fleet_list, fleet_status
 from autoskillit.cli._hooks import _claude_settings_path
 from autoskillit.cli._init_helpers import _prompt_recipe_choice
 from autoskillit.cli._mcp_names import detect_autoskillit_mcp_prefix
