+++ src/autoskillit/cli/fleet/_fleet_display.py
